requirements.txt
langchain>=0.1.0
langserve>=0.0.43
langchain-openai>=0.0.5
langchain-community>=0.0.13
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
python-dotenv>=1.0.0
httpx>=0.25.2
aiofiles>=23.2.1
sse-starlette>=1.6.5